requirements.txt
requests==2.31.0
feedparser==6.0.10
python-dotenv==1.0.0
jinja2==3.1.2
beautifulsoup4==4.12.0
httpx==0.25.0
meta-ai-api-tool-call==0.1.3
curl_cffi>=0.7.0
cloudscraper>=1.2.71
pycryptodome>=3.20.0
fake-useragent>=1.5.0
